Fix appimage_shall_not_be_integrated
The method didn't work with AppImages using a symlink in the AppDir root
as the desktop file. To check this, I changed Echo-no-integrate.

libappimage 1.0 also changed the contract of the method. The old
branch also ignored AppImages that just have NoDisplay set to true.

I agree that X-AppImage-Integrate is more explicit and using NoDisplay
isn't quite correct, however we can't change contracts of methods just
as we want.
Therefore we have to revert this change in the API.

Fixes #142.