Rework and update tests
Eliminates the dependency on Boost::filesystem and replaces it with
std::filesystem.

This commit further introduces a TemporaryDirectory class to replace and
improve the current temporary directories/files handling within the
tests with a RAII-style object. This class also makes the paths
(reasonably) predictable.

The commit also makes most objects used in the tests const where
appropriate.