Require some tools only when needed
Search and require the tools needed for the custom build of squashfuse
only in case we need to do the custom build of it, and not in all the
cases (i.e. when using a system version).

This avoids extra build time requirements when using a system version
of squashfuse.